fix(TD-JMATCHER-OFFSET): Change searchWindow GraphQL type from Int! to Long! (BL-186)
OTP GraphQL schema uses Long! for searchWindow but the query declaration used Int!,
causing a type mismatch runtime error on every /routes call. One-line fix to align
the variable declaration with the OTP server's expected type.

+++ src/services/otp-client.ts
@@ -44,7 +44,7 @@ const PLAN_JOURNEY_QUERY = `
     $fromLat: Float!, $fromLon: Float!,
     $toLat: Float!, $toLon: Float!,
     $date: String!, $time: String!,
-    $searchWindow: Int!
+    $searchWindow: Long!
   ) {
     plan(
       from: {lat: $fromLat, lon: $fromLon}
